+++ Frontend/src/components/Tile/Avatar/styled.js
