+++ Frontend/src/components/Tile/Footer/styled.js
