+++ Frontend/src/components/Tile/Heading/styled.js
