+++ Frontend/src/components/Tile/styled.js
